pyproject.toml
[project]
name = "rydercup-predictor-agent"
version = "0.1.0"
description = "Ryder Cup prediction agent using Google ADK and MCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agentops>=0.4.21",
    "google-adk>=0.1.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "mcp_servers"]
